+++ dionysus_app/persistence/database_functions.py
@@ -3,9 +3,11 @@
 from dionysus_app.persistence.database import Database
 from dionysus_app.persistence.databases.json import JSONDatabase
 from dionysus_app.persistence.databases.sqlite import SQLiteDatabase
+from dionysus_app.persistence.databases.sqlite_sqlalchemy import SQLiteSQLAlchemyDatabase
 
 database_backends = {'JSON': JSONDatabase,
-                     'SQLite': SQLiteDatabase
+                     'SQLite': SQLiteDatabase,
+                     'SQLiteSQLAlchemy': SQLiteSQLAlchemyDatabase,
                      }